feat/#20: 일정표 수정 사항 반영
